eventora-backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Migrate
Flask-Marshmallow
marshmallow-sqlalchemy
psycopg2-binary
Flask-JWT-Extended
Werkzeug
